% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_to_wd.R
\name{book}
\alias{book}
\title{Change to a particular working directory}
\usage{
book()
}
\value{
The last working directory, invisibly.
}
\description{
Move quickly to specified working directories on my system. These are named
the same way they are used via other programs on my computer  e.g. 
PowerShell.
}
\details{
\code{book} will go to the directory that has my personal study material.

Other functions will be added as necessary.
}
