% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winbuild.R
\name{check_on_winbuilder}
\alias{check_on_winbuilder}
\title{Run Package Checks on Win-Builder}
\usage{
check_on_winbuilder(pkg = ".")
}
\arguments{
\item{pkg}{The path to the package}
}
\description{
Automates the normal checks on the Win-Builder site
}
