% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{Check package installation status}
\usage{
check_package(pkg)
}
\arguments{
\item{pkg}{package being checked for.}
}
\value{
The function returns \code{TRUE} if the package is installed;
    \code{FALSE} if it is not.
}
\description{
\code{check_package} checks whether an R package is installed locally.
}
\examples{
## Not run
check_package("base")  # TRUE
}
