% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStuff.R
\name{create_and_edit}
\alias{create_and_edit}
\title{create_and_edit}
\usage{
create_and_edit(x = character())
}
\arguments{
\item{x}{a character vector of the filename(s)}
}
\description{
Create a brand-new file and immediately open it for editing.
}
