% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto_sandbox.R
\name{goto_sandbox}
\alias{goto_sandbox}
\title{Navigate to personal R sandbox}
\usage{
goto_sandbox(dir = NULL)
}
\arguments{
\item{dir}{A directory within the sandbox}
}
\description{
Navigate to personal R sandbox
}
