% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-functions.R
\name{install_R_my_way}
\alias{install_R_my_way}
\title{Customised Upgrading of R Installation}
\usage{
install_R_my_way()
}
\description{
Installing R using customised settings of
\code{\link[installr]{updateR}}
}
\details{
The R installation will be carried out without a recourse to 
any kind of interaction, with the exception of a possible prompt from
\emph{Windows Smartscreen}.
}
