% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_to_wd.R
\name{jump_to_wd}
\alias{jump_to_wd}
\title{Change to a particular working directory}
\usage{
jump_to_wd(x, home = NULL)
}
\arguments{
\item{x}{The destination directory}

\item{home}{The user's home directory; the starting point for the search}
}
\description{
This function is for those who are too lazy to type out a full path.
}
\details{
This function works best for the user who is vaguely familiar
    familiar with the directory tree structure of the work station, as
    there is a need to supply the exact name of the destination folder.
}

