% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-functions.R
\name{which_pkg_missing_description}
\alias{which_pkg_missing_description}
\title{Checks whether a (malformed) package is missing the 'DESCRIPTION' file}
\usage{
which_pkg_missing_description(lib)
}
\arguments{
\item{lib}{An R library directory}
}
\description{
Checks whether a (malformed) package is missing the 'DESCRIPTION' file
}
