% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStuff.R
\name{cheatsheet}
\alias{cheatsheet}
\title{Access/Open locally stored R cheatsheets}
\usage{
cheatsheet(doc = NULL)
}
\arguments{
\item{doc}{A character string of the filename, in part or whole. Regular
expressions are supported.}
}
\description{
Access/Open locally stored R cheatsheets
}
