% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_input.R
\name{collect_input}
\alias{collect_input}
\title{Prompt a User for and Receive Input During a Session}
\usage{
collect_input(prompt = "Enter: ", character.only = FALSE)
}
\arguments{
\item{prompt}{A message for prompting the user.}

\item{character.only}{Whether to strictly process as character. Defaults
to \code{FALSE}. It ensures that if the user provides a number, it is 
explicitly coerced to a numerical value. To enforce interpretation as a 
string requires it to be \code{TRUE}}
}
\description{
Collect input from within an active R session, whether during an
interactive session or while running R in batch mode.
}
