% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-functions.R
\name{old_pkg}
\alias{old_pkg}
\title{package-functions.R}
\usage{
old_pkg()
}
\description{
Save old R packages upon a new R installation.
}
\references{
https://www.datascienceriot.com/how-to-upgrade-r-without-losing-your-packages/kris/
}
